+++ Backend/ModelsWithDB/NewLoan/NL_PaypointTransactions.cs
@@ -8,11 +8,11 @@
     public class NL_PaypointTransactions {
 		[PK(true)]
         [DataMember]
-        public int PaypointTransactionID { get; set; }
+		public long PaypointTransactionID { get; set; }
 
         [FK("NL_Payments", "PaymentID")]
         [DataMember]
-        public int PaymentID { get; set; }
+		public long PaymentID { get; set; }
 
         [DataMember]
         public DateTime TransactionTime { get; set; }
